feat(browser compatibility): Generate a browser-compatible bundle for client & client full packages.
Generate a browser-comaptible bundle for client & client full packages using webpack. All packages
which rely on nodejs builtins are bundled, others are left as external.  Also allow createClient to
be configured with a host so that a private network running on localhost can be accessed from a
phone or emulator. This is all to aid with requested expo compatibility.